+++ src/main/resources/application.conf
@@ -14,7 +14,7 @@ akka {
     maximum-payload-bytes = 30000000 bytes
     log-remote-lifeclycle-events = off
     netty.tcp {
-      hostname = "192.168.1.107"
+      hostname = "10.100.1.48"//"192.168.1.107"
       message-frame-size =  30000000b
       send-buffer-size =  30000000b
       receive-buffer-size =  30000000b
@@ -23,7 +23,7 @@ akka {
   }
   cluster {
     seed-nodes = [
-      "akka.tcp://default@192.168.1.107:2550"
+      "akka.tcp://default@10.100.1.48:2550"
     ]
   }
 }
