+++ src/main/resources/camera3.conf
@@ -1,5 +1,5 @@
 akka.remote.netty.tcp{
-  hostname = "192.168.1.113"
+  hostname = "10.100.1.160"//"192.168.1.113"
   port = 2551
 }
 akka.cluster.roles = ["camera"]
@@ -8,7 +8,7 @@ camera {
   name = "raspberry"
   periodicPhotos {
     active = true
-    period = 5s
+    period = 2s
   }
   port = 9091
 }
