+++ src/main/scala/utils/SwaggerConfig.scala
@@ -1,18 +1,13 @@
 package utils
 
-import akka.actor.ActorSystem
-import akka.stream.ActorMaterializer
-import com.github.swagger.akka.{HasActorSystem, SwaggerHttpService}
-import scala.reflect.runtime.universe._
+import com.github.swagger.akka.SwaggerHttpService
 import endpoint.PhotoEndpoint
 
 
 /**
   * Created by Alfonso on 06/02/2016.
   */
-class SwaggerConfig(system: ActorSystem) extends SwaggerHttpService with HasActorSystem {
-  implicit val actorSystem: ActorSystem = system
-  implicit val materializer: ActorMaterializer = ActorMaterializer()
-  override val apiTypes: Seq[scala.reflect.runtime.universe.Type] = Seq(typeOf[PhotoEndpoint])
-  //override here to change the SwaggerHttpService methods config
+class SwaggerConfig() extends SwaggerHttpService {
+  override val apiClasses = Set(classOf[PhotoEndpoint])
+  override val host = "localhost:8080"
 }
